updated API

+++ server.js
@@ -1,10 +1,11 @@
 // Built-in Node.js modules
-let fs = require('fs');
-let path = require('path');
-
+const fs = require('fs');
+const path = require('path');
 // NPM modules
 let express = require('express');
 let sqlite3 = require('sqlite3');
+const fetch = (...args) => import('node-fetch').then(({default: fetch}) => fetch(...args));
+
 
 
 let db_filename = path.join(__dirname, 'db', 'stpaul_crime.sqlite3');
@@ -24,6 +25,25 @@ let db = new sqlite3.Database(db_filename, sqlite3.OPEN_READWRITE, (err) => {
     else {
         console.log('Now connected to ' + path.basename(db_filename));
     }
+});
+
+// TODO: change this to a GET:// with a URL encoded parameter, maybe
+app.post('/lookup', async (req, res) => {
+    const address = req.body.address;
+    if (!address) {
+        res.status(400).type('json').send({error: 'Validation Error: address is required'});
+    }
+
+    console.log('Looking up the address for ',address)
+    const response = await fetch(`https://nominatim.openstreetmap.org/search?q=${address}&format=json`, {
+        method: 'GET',
+        headers: {
+            ['content-type']: 'application/json'
+        }
+    });
+    const data = await response.json()
+    console.log('Got the response', data)
+    res.status(200).type('json').send(data);
 });
 
 
@@ -78,6 +98,11 @@ app.get('/neighborhoods', (req, res) => {
       });
 });
 
+function extractListFromQueryParam(val) {
+    return (Array.isArray(val) ? val : [val])
+        .flatMap(el => el.split(','))
+}
+
 // GET request handler for crime incidents
 app.get('/incidents', (req, res) => {
     console.log(req.query); // query object (key-value pairs after the ? in the url)
@@ -90,87 +115,90 @@ app.get('/incidents', (req, res) => {
         let j = 0;
         let and = ' AND';
         let where = ' WHERE';
-        for(k in req.query){
+        for(k in req.query) {
             if(k == 'code'){
                 if(i==0){
-                    sql = sql + where;
+                    sql += where;
                 }
                 else{
-                    sql = sql + and;
+                    sql += and;
                 }
                 sql = sql + " code IN (?)";
-                p.push(req.query.code);
-                i = i + 1;
+                // TODO: we might want to be returning `...flatMap().join(,)` from extractListFromQueryParam
+                // if this errors out that's probably why
+                p.push(extractListFromQueryParam(req.query.code));
+                i++;
             }
             else if(k == 'start_date'){
                 if(i==0){
-                    sql = sql + where;
+                    sql += where;
                 }
                 else{
-                    sql = sql + and;
+                    sql += and;
                 }
                 sql = sql + " date > ?";
                 p.push(req.query.start_date);
-                i = i + 1;
+                i++;
             }
             else if(k == 'end_date'){
                 if(i==0){
-                    sql = sql + where;
+                    sql += where;
                 }
                 else{
-                    sql = sql + and;
+                    sql += and;
                 }
-                sql = sql + " WHERE date < ?";
+                sql = sql + " date < ?";
                 p.push(req.query.end_date);
-                i = i + 1;
+                i++;
             }
             else if(k == 'grid'){
                 if(i==0){
-                    sql = sql + where;
+                    sql += where;
                 }
                 else{
-                    sql = sql + and;
+                    sql += and;
                 }
-                sql = sql + " WHERE police_grid IN (?)";
-                p.push(req.query.grid);
-                i = i + 1;
+                sql = sql + " police_grid IN (?)";
+                p.push(extractListFromQueryParam(req.query.grid));
+                i++;
             }
             else if(k == 'neighborhood'){
                 if(i==0){
-                    sql = sql + where;
+                    sql += where;
                 }
                 else{
-                    sql = sql + and;
+                    sql += and;
                 }
-                sql = sql + " WHERE police_grid IN (?)";
-                p.push(req.query.grid);
-                i = i + 1;
+                sql = sql + " neighborhood_number IN (?)";
+                p.push(extractListFromQueryParam(req.query.neighborhood_number));
+                i++;
             }
-            else if(k == 'limit'){
-                sql = sql + "ORDER BY date DESC LIMIT (?)"; 
-                p.push(req.query.limit); 
-                i = i + 1;
+            else if(k == 'limit') {
+                sql = sql + " ORDER BY date DESC LIMIT (?)";
+                p.push(req.query.limit);
+                i++;
                 j = j + 1;
             }
-        }
-        if(j == 0){
-            sql = sql + "ORDER BY date DESC LIMIT 1000";
+
+            if(j == 0){
+                sql = sql + " ORDER BY date DESC LIMIT 1000";
+            }
         }
     }
-    
+        
     db.all(sql, p, (err, rows) => {
         var mydata = []; //once this was inside the db method, the assignment became synchcronous
         if (err) {
-          throw err;
+            rs.status(500).type('json').send({ error: 'INTERNAL SERVER ERROR'})
         }
         rows.forEach((row) => {
-          console.log(row.date);
-          mydata.push(row);
+            console.log(row.date);
+            mydata.push(row);
         });
-        res.status(200).type('json').send(mydata); 
         //this returns the response inside the db method
         //so that the data is synchronized
-      });
+        res.status(200).type('json').send(mydata); 
+    });
 });
 
 // PUT request handler for new crime incident
@@ -191,16 +219,26 @@ app.put('/new-incident', (req, res) => {
 });
 
 // DELETE request handler for new crime incident
-app.delete('/delete-incident', (req, res) => {
+app.delete('/remove-incident', (req, res) => {
     console.log(req.body); // uploaded data
-    let sql = "DELETE FROM Incidents WHERE case_number = (?)";
-    databaseRun(sql, req.body.case_number)
-    .then(() => {
-        res.status(200).type('txt').send('OK');
-    })
-    .catch((err) => {
-        res.status(500).type('txt').send('NOT OK')
-    })
+    databaseSelect("SELECT * from Incidents WHERE case_number = (?)", (err, res)=>{
+        if (err) {
+            //TODO:  handle error
+        } else {
+            if (res) {
+                let sql = "DELETE FROM Incidents WHERE case_number = (?)";
+                databaseRun(sql, req.body.case_number).then(() => {
+                    res.status(200).type('txt').send('OK');
+                })
+                .catch((err) => {
+                    res.status(500).type('txt').send('NOT OK')
+                })
+            } else {
+                // prof complaint
+                //TODO: handle 204 NO CONTENT
+            }
+        }
+    });
 });
